requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
python-docx
pdfplumber
pydantic
